+++ anomalib/deploy/__init__.py
@@ -3,7 +3,7 @@
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .export import OutputFormat, export, get_model_metadata
+from .export import ExportFormat, export, get_model_metadata
 from .inferencers import Inferencer, OpenVINOInferencer, TorchInferencer
 
-__all__ = ["OutputFormat", "Inferencer", "OpenVINOInferencer", "TorchInferencer", "export", "get_model_metadata"]
+__all__ = ["ExportFormat", "Inferencer", "OpenVINOInferencer", "TorchInferencer", "export", "get_model_metadata"]