+++ tests/pre_merge/utils/callbacks/export_callback/test_export.py
@@ -5,7 +5,7 @@ import pytest
 import pytorch_lightning as pl
 from pytorch_lightning.callbacks.early_stopping import EarlyStopping
 
-from anomalib.deploy import OutputFormat
+from anomalib.deploy import ExportFormat
 from anomalib.utils.callbacks.export import ExportCallback
 from tests.helpers.config import get_test_configurable_parameters
 from tests.pre_merge.utils.callbacks.export_callback.dummy_lightning_model import (
@@ -15,10 +15,10 @@ from tests.pre_merge.utils.callbacks.export_callback.dummy_lightning_model impor
 
 
 @pytest.mark.parametrize(
-    "output_format",
-    [OutputFormat.OPENVINO, OutputFormat.ONNX],
+    "export_format",
+    [ExportFormat.OPENVINO, ExportFormat.ONNX],
 )
-def test_export_callback(output_format):
+def test_export_callback(export_format):
     """Tests if an optimized model is created."""
 
     config = get_test_configurable_parameters(
@@ -33,7 +33,7 @@ def test_export_callback(output_format):
                 input_size=config.data.init_args.image_size,
                 dirpath=os.path.join(tmp_dir),
                 filename="model",
-                output_format=output_format,
+                export_format=export_format,
             ),
             EarlyStopping(monitor=config.model.init_args.metric),
         ]
@@ -48,9 +48,9 @@ def test_export_callback(output_format):
         )
         trainer.fit(model, datamodule=datamodule)
 
-        if output_format == OutputFormat.OPENVINO:
+        if export_format == ExportFormat.OPENVINO:
             assert os.path.exists(os.path.join(tmp_dir, "openvino/model.bin")), "Failed to generate OpenVINO model"
-        elif output_format == OutputFormat.ONNX:
+        elif export_format == ExportFormat.ONNX:
             assert os.path.exists(os.path.join(tmp_dir, "onnx/model.onnx")), "Failed to generate ONNX model"
         else:
-            raise ValueError(f"Unknown format {output_format}. Supported modes: onnx or openvino.")
+            raise ValueError(f"Unknown format {export_format}. Supported modes: onnx or openvino.")